se reemplaza hash viejo con el nuevo
